+++ src/shims-vue.d.ts
@@ -9,3 +9,4 @@ declare module "mgrs";
 declare module "ol-ext/layer/GeoImage";
 declare module "ol-ext/source/GeoImage";
 declare module "ol-ext/interaction/Transform";
+declare module "ol-ext/source/DayNight";