+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 March 2024
 
+- Add day/night terminator to the map.
 - Add unit path timestamp toggle.
 - Add a unit path panel to the main toolbar.